# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "IdeaPress",
    "version": "1.5.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/projects/{project_id}/export": {
      "get": {
        "description": "Render the committed project and return it, without writing anything.\n\nRaises:\n    ExportFailed: The format is not one of the three shipped at 1.0.",
        "operationId": "get_export_api_v1_projects__project_id__export_get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "format",
            "required": false,
            "schema": {
              "default": "markdown",
              "title": "Format",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Export",
        "tags": [
          "export"
        ]
      },
      "post": {
        "description": "Write the export into the project's directory and return the artifact record.",
        "operationId": "post_export_api_v1_projects__project_id__export_post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "format",
            "required": false,
            "schema": {
              "default": "markdown",
              "title": "Format",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Post Export",
        "tags": [
          "export"
        ]
      }
    },
    "/api/v1/projects/{project_id}/plan": {
      "get": {
        "description": "The compiled requirements, each with its source quote and checks, and the unit plan.\n\nRaises:\n    ProjectNotFound: No such project.",
        "operationId": "get_plan_api_v1_projects__project_id__plan_get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Plan",
        "tags": [
          "plan"
        ]
      },
      "post": {
        "description": "Compile requirements and build the unit plan. Returns the task.\n\nRaises:\n    ProjectNotFound: No such project.\n    StageAlreadyRunning: A stage is already running for this project.",
        "operationId": "post_plan_api_v1_projects__project_id__plan_post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Post Plan",
        "tags": [
          "stages"
        ]
      }
    },
...
    "/api/v1/projects/{project_id}/research": {
      "get": {
        "description": "Where a fetch may go, and every note and tool call the research stage left (row WP5).\n\nThe stage starts at ``POST /projects/{id}/stages/research/run``. This is what it may reach\nbefore anyone starts it (spec \u00a714: a click that may fetch says where it can go) and what it\nrecorded afterwards, each call with the egress decision it ran under.\n\nRaises:\n    ProjectNotFound: No such project.",
        "operationId": "get_research_api_v1_projects__project_id__research_get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Research",
        "tags": [
          "projects"
        ]
      }
    },
...
    "/api/v1/projects/{project_id}/workspace": {
      "get": {
        "description": "The workspace page's own view of one unit, over the API (row WP5).\n\nAssembled by the same ``workspace_view`` the page renders: the navigator, the unit's content\nand provenance, its pause reason with IdeaPress's remedy, its coverage summary, the diff\nagainst ``compare`` when asked, the backend's recorded egress facts, the project's cost, the\nrunning stage and where research may fetch.\n\nArgs:\n    request: The request.\n    project_id: The project.\n    unit: Which unit; the first one when empty or unknown.\n    compare: A version to diff the current one against.\n\nRaises:\n    ProjectNotFound: No such project.",
        "operationId": "get_workspace_api_v1_projects__project_id__workspace_get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "unit",
            "required": false,
            "schema": {
              "default": "",
              "title": "Unit",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "compare",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Compare"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Workspace",
        "tags": [
          "workspace"
        ]
      }
    },
...
  }
}